+++ packages/compiler/src/init/scaffold.ts
@@ -1,18 +1,17 @@
-import { stringify } from "yaml";
-import { TypeSpecConfigFilename } from "../config/config-loader.js";
-import { TypeSpecRawConfig } from "../config/types.js";
-import { formatTypeSpec } from "../core/formatter.js";
+import type { TypeSpecRawConfig } from "../config/types.js";
 import { getDirectoryPath, joinPaths } from "../core/path-utils.js";
-import { CompilerHost } from "../core/types.js";
-import { PackageJson } from "../types/package-json.js";
+import type { SystemHost } from "../core/types.js";
+import type { PackageJson } from "../types/package-json.js";
 import { readUrlOrPath, resolveRelativeUrlOrPath } from "../utils/misc.js";
-import { FileTemplatingContext, createFileTemplatingContext, render } from "./file-templating.js";
-import {
+import { createFileTemplatingContext, render } from "./file-templating.js";
+import type {
   InitTemplate,
   InitTemplateFile,
   InitTemplateLibrary,
   InitTemplateLibrarySpec,
 } from "./init-template.js";
+
+export const TypeSpecConfigFilename = "tspconfig.yaml";
 
 export interface ScaffoldingConfig {
   /** Template used to resolve that config */
@@ -89,7 +88,7 @@ export function makeScaffoldingConfig(
  * @param host
  * @param config
  */
-export async function scaffoldNewProject(host: CompilerHost, config: ScaffoldingConfig) {
+export async function scaffoldNewProject(host: SystemHost, config: ScaffoldingConfig) {
   await host.mkdirp(config.directory);
   await writePackageJson(host, config);
   await writeConfig(host, config);
@@ -107,7 +106,7 @@ export function isFileSkipGeneration(fileName: string, files: InitTemplateFile[]
   return false;
 }
 
-async function writePackageJson(host: CompilerHost, config: ScaffoldingConfig) {
+async function writePackageJson(host: SystemHost, config: ScaffoldingConfig) {
   if (isFileSkipGeneration("package.json", config.template.files ?? [])) {
     return;
   }
@@ -138,10 +137,7 @@ async function writePackageJson(host: CompilerHost, config: ScaffoldingConfig) {
     private: true,
   };
 
-  return host.writeFile(
-    joinPaths(config.directory, "package.json"),
-    JSON.stringify(packageJson, null, 2),
-  );
+  return host.writeFile("package.json", JSON.stringify(packageJson, null, 2));
 }
 
 const placeholderConfig = `
@@ -162,7 +158,7 @@ const placeholderConfig = `
 # warn-as-error: true                           # Treat warnings as errors
 # output-dir: "{project-root}/_generated"       # Configure the base output directory for all emitters
 `.trim();
-async function writeConfig(host: CompilerHost, config: ScaffoldingConfig) {
+async function writeConfig(host: SystemHost, config: ScaffoldingConfig) {
   if (isFileSkipGeneration(TypeSpecConfigFilename, config.template.files ?? [])) {
     return;
   }
@@ -180,11 +176,11 @@ async function writeConfig(host: CompilerHost, config: ScaffoldingConfig) {
       Object.entries(config.emitters).map(([key, emitter]) => [key, emitter.options]),
     );
   }
-  const content = rawConfig ? stringify(rawConfig) : placeholderConfig;
-  return host.writeFile(joinPaths(config.directory, TypeSpecConfigFilename), content);
+  const content = rawConfig ? "" : placeholderConfig;
+  return host.writeFile(TypeSpecConfigFilename, content);
 }
 
-async function writeMain(host: CompilerHost, config: ScaffoldingConfig) {
+async function writeMain(host: SystemHost, config: ScaffoldingConfig) {
   if (isFileSkipGeneration("main.tsp", config.template.files ?? [])) {
     return;
   }
@@ -197,7 +193,7 @@ async function writeMain(host: CompilerHost, config: ScaffoldingConfig) {
   const lines = [...config.libraries.map((x) => `import "${x.name}";`), ""];
   const content = lines.join("\n");
 
-  return host.writeFile(joinPaths(config.directory, "main.tsp"), await formatTypeSpec(content));
+  // return host.writeFile(joinPaths(config.directory, "main.tsp"), await formatTypeSpec(content));
 }
 
 const defaultGitignore = `
@@ -211,15 +207,15 @@ dist/
 # Dependency directories
 node_modules/
 `.trim();
-async function writeGitignore(host: CompilerHost, config: ScaffoldingConfig) {
+async function writeGitignore(host: SystemHost, config: ScaffoldingConfig) {
   if (!config.includeGitignore || isFileSkipGeneration(".gitignore", config.template.files ?? [])) {
     return;
   }
 
-  return host.writeFile(joinPaths(config.directory, ".gitignore"), defaultGitignore);
+  return host.writeFile(".gitignore", defaultGitignore);
 }
 
-async function writeFiles(host: CompilerHost, config: ScaffoldingConfig) {
+async function writeFiles(host: SystemHost, config: ScaffoldingConfig) {
   const templateContext = createFileTemplatingContext(config);
   if (!config.template.files) {
     return;
@@ -232,9 +228,9 @@ async function writeFiles(host: CompilerHost, config: ScaffoldingConfig) {
 }
 
 async function writeFile(
-  host: CompilerHost,
+  host: SystemHost,
   config: ScaffoldingConfig,
-  context: FileTemplatingContext,
+  context: any,
   file: InitTemplateFile,
 ) {
   const baseDir = config.baseUri + "/";
