+++ packages/compiler/src/internals/init.ts
@@ -1,0 +1,3 @@
+// export { getTypeSpecCoreTemplates } from "../init/core-templates.js";
+export { NodeSystemHost } from "../core/node-system-host.js";
+export { scaffoldNewProject } from "../init/scaffold.js";
