+++ packages/compiler/src/server/compile-service.ts
@@ -21,7 +21,8 @@ import {
 } from "../core/index.js";
 import { builtInLinterLibraryName, builtInLinterRule_UnusedUsing } from "../core/linter.js";
 import { compile as compileProgram } from "../core/program.js";
-import { doIO, loadFile, resolveTspMain } from "../utils/misc.js";
+import { doIO, loadFile } from "../utils/io.js";
+import { resolveTspMain } from "../utils/misc.js";
 import { serverOptions } from "./constants.js";
 import { FileService } from "./file-service.js";
 import { FileSystemCache } from "./file-system-cache.js";
