+++ packages/typespec-vscode/package.json
@@ -240,12 +240,13 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "~28.0.2",
+    "@rollup/plugin-json": "~6.1.0",
     "@rollup/plugin-node-resolve": "~16.0.0",
     "@rollup/plugin-typescript": "~12.1.0",
     "@types/mocha": "^10.0.9",
     "@types/node": "~22.10.10",
-    "@types/vscode": "~1.96.0",
     "@types/semver": "^7.5.8",
+    "@types/vscode": "~1.96.0",
     "@typespec/compiler": "workspace:~",
     "@typespec/internal-build-utils": "workspace:~",
     "@vitest/coverage-v8": "^3.0.4",
